chore(Mappie): add simple-import-sort plugin, fix import placement, split styles file
