+++ src/components/FilterItem/FilterItem.styled.ts
@@ -1,10 +1,7 @@
 import styled from 'styled-components';
-import {
-  flexGap,
-  fontWeights,
-  BaseIconStyle,
-  colors,
-} from '../../common/common.styled';
+
+import { colors, fontWeights } from '../../constants/theme';
+import { BaseIconStyle, flexGap } from '../../styles/BaseStyle';
 
 export const Wrapper = styled.div<{ selected: boolean }>`
   ${flexGap('15px')};
