+++ src/components/SearchInput/SearchInput.styled.ts
@@ -1,12 +1,8 @@
 import styled from 'styled-components';
+
 import IconSearch from '../../assets/icons/IconSearch.svg?react';
-import {
-  colors,
-  border,
-  fontWeights,
-  BaseInput,
-  BaseIconStyle,
-} from '../../common/common.styled';
+import { border, colors, fontWeights } from '../../constants/theme';
+import { BaseIconStyle, BaseInput } from '../../styles/BaseStyle';
 
 export const InputWrapper = styled.div`
   position: relative;
