+++ src/components/SearchResultsPanel/SearchResultsPanel.styled.ts
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
-import { BasePanel, colors } from '../../common/common.styled';
+
+import { colors } from '../../constants/theme';
+import { BasePanel } from '../../styles/BaseStyle';
 
 export const Panel = styled(BasePanel)`
   width: 400px;
