+++ src/components/Sidebar/Sidebar.styled.ts
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
-import { colors, flexColumn } from '../../common/common.styled';
+
+import { colors } from '../../constants/theme';
+import { flexColumn } from '../../styles/BaseStyle';
 
 export const SidebarWrapper = styled.div`
   width: 110px;
