+++ src/constants/theme.ts
@@ -1,0 +1,49 @@
+export const colors = {
+  primaryBlue: '#5E7BC7',
+  accentRed: '#C75E5E',
+  darkGrey: '#373737',
+  mediumGrey: '#808080',
+  lightGreyBorder: '#C4C4C4',
+  white: '#FFFFFF',
+  black: '#000000',
+  blackShadow: '#0000001a',
+  errorRed: '#FF0000',
+  darkText: '#333',
+  lightBorder: '#ccc',
+  googleRed: '#FA4747',
+  googleBorder: '#DADCE0',
+  monoRed: '#C75E5E',
+  primaryBlueAlpha: 'rgba(94, 123, 199, 0.5)',
+  darkShadowAlpha: '#0000003d',
+  lighterBorder: '#ddd',
+  transparentBlackBorder: '#00000033',
+};
+
+export const fontFamilies = {
+  primary: 'Mont, sans-serif',
+  monospace: 'monospace',
+};
+
+export const fontWeights = {
+  extraBold: 800,
+  bold: 700,
+  medium: 600,
+  regular: 400,
+};
+
+export const borderRadius = {
+  large: '10px',
+  medium: '6px',
+  small: '5px',
+};
+
+export const border = {
+  default: `3px solid ${colors.lightGreyBorder}`,
+  none: 'none',
+  input: `1px solid ${colors.lightBorder}`,
+};
+
+export const shadows = {
+  default: `1px 0px 20px 0px ${colors.blackShadow}`,
+  form: `1px 0px 30px 0px ${colors.darkShadowAlpha}`,
+};
