+++ src/main.tsx
@@ -1,21 +1,19 @@
-import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import App from './App.tsx';
-import { GlobalStyle } from './styles/GlobalStyle.ts';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
-import { store } from './store/store.ts';
 import { BrowserRouter } from 'react-router-dom';
 
+import App from './App.tsx';
+import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
+import { store } from './store/store.ts';
+import { GlobalStyle } from './styles/GlobalStyle.ts';
+
 createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <Provider store={store}>
-      <GlobalStyle />
-      <ErrorBoundary>
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
-      </ErrorBoundary>
-    </Provider>
-  </StrictMode>
+  <Provider store={store}>
+    <GlobalStyle />
+    <ErrorBoundary>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </ErrorBoundary>
+  </Provider>
 );
