+++ src/pages/RegisterPage/RegisterPage.tsx
@@ -1,14 +1,15 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+
 import { useAuth } from '../../hooks/useAuth';
 import { isValidEmail, isValidPassword } from '../../utils/validation';
 import {
   Button,
   ErrorMsg,
+  Form,
   InfoMsg,
   Input,
   Switch,
-  Form,
 } from './RegisterPage.styled';
 
 function RegisterPage() {
