+++ src/routes/RequireAuth.tsx
@@ -1,5 +1,6 @@
-import { Navigate, useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
+import { Navigate, useLocation } from 'react-router-dom';
+
 import type { RootState } from '../store/store';
 
 export function RequireAuth({ children }: { children: React.ReactElement }) {
