+++ src/slices/poiSlice.ts
@@ -1,10 +1,11 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
+import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
+
 import type {
+  IFeatureItem,
+  IOverpassPOIItem,
   IPOI,
   IPOIState,
-  IOverpassPOIItem,
-  IFeatureItem,
 } from '../types/interfaces';
 
 const API_KEY = import.meta.env.VITE_OPENTRIPMAP_API_KEY;
