+++ src/slices/userSlice.ts
@@ -1,6 +1,7 @@
-import type { IUserSlice } from '../types/interfaces';
-import { createSlice } from '@reduxjs/toolkit';
 import type { PayloadAction } from '@reduxjs/toolkit';
+import { createSlice } from '@reduxjs/toolkit';
+
+import type { IUserSlice } from '../types/interfaces';
 
 const initialState: IUserSlice = {
   email: null,
