+++ src/store/store.ts
@@ -1,9 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
-import userReducer from '../slices/userSlice';
-import poiReducer from '../slices/poiSlice';
+
 import favoriteReducer from '../slices/favoriteSlice';
-import userLocationReducer from '../slices/userLocationSlice';
+import poiReducer from '../slices/poiSlice';
 import routeRedicer from '../slices/routeSlice';
+import userLocationReducer from '../slices/userLocationSlice';
+import userReducer from '../slices/userSlice';
 
 export const store = configureStore({
   reducer: {
